Moved spoiler log generation to RandomizerLib and added server side spoiler file
